// src/model/account.ts
export interface Account {
    id: number
    name: string
    channel_info: string
    category: string
    posts: number
    followers: number
    avg_likes: number
    eng_rate: string
    email: string
    phone: string
    display_posts: string
    display_avg_likes: string
    display_followers: string
}